requirements.txt
fpdf==1.7.2
lxml==4.1.1
pkg-resources==0.0.0
PyPDF2==1.26.0
python-docx==0.8.6
